Introduce base class for structures and make tile and structure separate object
This way, the need to check whether a method is defined is eliminated.

Tile and structure should be separate entities because tile has geometric
properties that are not affected by structure on it.  And we need to
clean the structure without touching the tile underneath it.
It introduces an extra layer of objects, which could be more complex
for some people, but it's more closer to actual data model.