Fix previous commit didn't actually work on Edge but IE11
The acceptable data type for DataTransfer.setData() is different among
IE11 and Edge (EVEN THOUGH IE11 ought to support HTML5), so we need to
change type string according to the browser type.

It works on both IE11 and Edge except compatibility mode in Edge.
For some reason, Edge's emulation for IE11 is not completely compatible
with IE11.