Prevent drag & dropping images in inventory
HTML img element's draggable attribute is true by default, but actually
we do not want it to be dragged, we want the containing div to be dragged.
So we have to explicity turn off draggable attribute for images in
inventory list, no matter list or icon mode.

Also set some buttons and title to ignore dragging and selection (which in
turn enables dragging).