Add capacity and type limit to inventory owners
We knowingly used direct reference to hash object which represents
inventory itself (like selectedInventory), but it won't be good practice
because we cannot call methods on that inventory without reference to its
owner.  We designed in this way because the player is not object-like at
all, but now that the player has the same interface as Containers, we can
use methods like addItem or removeItem without knowing if the object is
actually a Container or a player.

The implementation is a bit like multiple inheritance in a way that the
player is not a direct subclass of Container but shares certain methods and
the data member 'inventory'.  Normally we would do this by mix-in, but we
didn't define a class for the player (the player has methods as fields)
since the player is singleton class, there's no point defining another
object representing its class.  (Well, it's not like totally pointless, but
we're lazy here.)  If inventory management needs additional methods,
we may move forward to full-fledged OO class definition.