Force DropItems to be aligned on center of transport belts
Just like Factorio, transport belts should have abilities to align items on them,
otherwise it would be a mess, spilling items over after a few turns.

The difference between this game and Factorio is that Factorio has two lanes
on a belt, but it would complicate factory design too much.
Also, item transition movements on curves and merges are not optimized as
Factorio.  We're using JavaScript, there is a limit to optimize anyway.