An attempt to reduce laggy behavior of MS Edge
Edge seems to take much more time to process mousemove event on tile elements
than Chrome or Firefox.  The reason is unclear, but it's updating info panel
unnecessarily frequently, so we try to reduce it and see if it helps.

There are still noticeable micro-pauses when the mouse cursor crosses the
border of tiles, but it's not as frequent as every mouse move.
It's certainly less frustrating than before, based on my subjective
judgment.