Update the body element's width style so that contents fit inside it
Devices with small screens like smartphones have trouble displaying the
game contents because the body element has smaller width than its
contents.  The body element's shape won't be updated as the contents
are drawn, because everything inside it is located with position: relative.
So we need to update it manually.

We haven't been aware of this problem because it wasn't a problem in
bigger monitors such as computers.