Synchronize animation currentTime when a belt is placed, using Web Animations API
It is kind of difficult decision whether to use Web Animations API because
it is still a draft as of 2020, and browsers have varying levels of support.

See:

https://caniuse.com/web-animation
https://developer.mozilla.org/en-US/docs/Web/API/Web_Animations_API/Using_the_Web_Animations_API

It will definitely improve rendering performance, but if the browser is
new enough to support Web Animations API, it should support canvas too,
in which case we can just use WebGL to optimize rendering.